requirements.txt
requests
feedparser
trafilatura
readability-lxml
python-frontmatter
pyyaml
openai
anthropic
python-dotenv
